Addressed review suggestions
