+++ servicetalk-http-api/src/main/java/io/servicetalk/http/api/StrategyInfluencerAwareConversions.java
@@ -19,57 +19,13 @@ import java.util.function.Predicate;
 
 import static java.util.Objects.requireNonNull;
 
-/**
- * Utility for converting filter factories to conditional factories which are applied when provided
- * {@link Predicate} returns {@code true}. The returned implementations comply to the
- * {@link HttpExecutionStrategyInfluencer} hint interface and apply chosen
- * {@link io.servicetalk.transport.api.ExecutionStrategy} when invoked.
- */
-public final class StrategyInfluencerAwareConversions {
+final class StrategyInfluencerAwareConversions {
 
     private StrategyInfluencerAwareConversions() {
         // No instances.
     }
 
-    /**
-     * Converts a {@link StreamingHttpServiceFilterFactory} to one that is conditional upon the provided
-     * {@link Predicate}.
-     * @param predicate When {@code true} the filter logic is applied.
-     * @param original Original {@link StreamingHttpServiceFilterFactory}.
-     * @return {@link StreamingHttpServiceFilterFactory} which maintains the {@link HttpExecutionStrategyInfluencer}
-     * interface if the original implements it.
-     */
-    public static StreamingHttpServiceFilterFactory toConditionalServiceFilterFactory(
-            final Predicate<StreamingHttpRequest> predicate, final StreamingHttpServiceFilterFactory original) {
-        requireNonNull(predicate);
-        requireNonNull(original);
-
-        if (original instanceof HttpExecutionStrategyInfluencer) {
-            HttpExecutionStrategyInfluencer influencer = (HttpExecutionStrategyInfluencer) original;
-            return new StrategyInfluencingStreamingServiceFilterFactory() {
-                @Override
-                public StreamingHttpServiceFilter create(final StreamingHttpService service) {
-                    return new ConditionalHttpServiceFilter(predicate, original.create(service), service);
-                }
-
-                @Override
-                public HttpExecutionStrategy influenceStrategy(final HttpExecutionStrategy strategy) {
-                    return influencer.influenceStrategy(strategy);
-                }
-            };
-        }
-        return service -> new ConditionalHttpServiceFilter(predicate, original.create(service), service);
-    }
-
-    /**
-     * Converts a {@link StreamingHttpConnectionFilterFactory} to one that is conditional upon the provided
-     * {@link Predicate}.
-     * @param predicate When {@code true} the filter logic is applied.
-     * @param original Original {@link StreamingHttpConnectionFilterFactory}.
-     * @return {@link StreamingHttpConnectionFilterFactory} which maintains the {@link HttpExecutionStrategyInfluencer}
-     * interface if the original implements it.
-     */
-    public static StreamingHttpConnectionFilterFactory toConditionalConnectionFilterFactory(
+    static StreamingHttpConnectionFilterFactory toConditionalConnectionFilterFactory(
             final Predicate<StreamingHttpRequest> predicate, final StreamingHttpConnectionFilterFactory original) {
         requireNonNull(predicate);
         requireNonNull(original);
@@ -91,15 +47,7 @@ public final class StrategyInfluencerAwareConversions {
         return connection -> new ConditionalHttpConnectionFilter(predicate, original.create(connection), connection);
     }
 
-    /**
-     * Converts a {@link StreamingHttpClientFilterFactory} to one that is conditional upon the provided
-     * {@link Predicate}.
-     * @param predicate When {@code true} the filter logic is applied.
-     * @param original Original {@link StreamingHttpClientFilterFactory}.
-     * @return {@link StreamingHttpClientFilterFactory} which maintains the {@link HttpExecutionStrategyInfluencer}
-     * interface if the original implements it.
-     */
-    public static StreamingHttpClientFilterFactory toConditionalClientFilterFactory(
+    static StreamingHttpClientFilterFactory toConditionalClientFilterFactory(
             final Predicate<StreamingHttpRequest> predicate, final StreamingHttpClientFilterFactory original) {
         requireNonNull(predicate);
         requireNonNull(original);
@@ -119,10 +67,6 @@ public final class StrategyInfluencerAwareConversions {
             };
         }
         return client -> new ConditionalHttpClientFilter(predicate, original.create(client), client);
-    }
-
-    interface StrategyInfluencingStreamingServiceFilterFactory
-            extends StreamingHttpServiceFilterFactory, HttpExecutionStrategyInfluencer {
     }
 
     interface StrategyInfluencingStreamingConnectionFilterFactory
